+++ .gitignore
@@ -17,7 +17,7 @@ venv/
 
 *.idea
 
-*Pipfile.lock
+
 # *media
 *.DS_Store
 /.vscode
