+++ backend/views.py
@@ -12,7 +12,6 @@ from rest_framework.decorators import api_view, parser_classes
 from rest_framework import status, generics
 import requests
 import time
-from .utils import send_centrifugo_data
 from .db import *
 from rest_framework.parsers import MultiPartParser, FormParser
 from rest_framework.views import (
@@ -20,15 +19,20 @@ from rest_framework.views import (
     exception_handler,
 )
 from django.core.files.storage import default_storage
+
 # Import Read Write function to Zuri Core
 from .resmodels import *
 from .serializers import *
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
+from .utils import SendNotificationThread
 from datetime import datetime
 import datetime as datetimemodule
+
+
 from .centrifugo_handler import centrifugo_client
 from rest_framework.pagination import PageNumberPagination
+
 from .decorators import db_init_with_credentials
 
 
@@ -727,7 +731,7 @@ def pinned_message(request, message_id):
         return Response(status=status.HTTP_503_SERVICE_UNAVAILABLE)
     if message_id in pin:
         pin.remove(message_id)
-        data = {"message_id": message_id, "pinned": pin, "Event": "unpin_message"} # this event key is in capslock
+        data = {"message_id": message_id, "pinned": pin, "Event": "unpin_message"}
         response = DB.update("dm_rooms", room_id, {"pinned": pin})
         # room = DB.read("dm_rooms", {"id": room_id})
         if response["status"] == 200:
@@ -865,30 +869,28 @@ def user_profile(request, org_id, member_id):
     responses={400: "Error: Bad Request"},
 )
 @api_view(["POST"])
-def create_reminder(request):
+@db_init_with_credentials
+def remind_message(request):
     """
         This is used to remind a user about a  message
         Your body request should have the format
         {
-        "message_id": "6146ea68845b436ea04d107d",
+        "mesage_id": "33",
         "current_date": "Tue, 22 Nov 2011 06:00:00 GMT",
-        "scheduled_date":"Tue, 22 Nov 2011 06:10:00 GMT",
-        "notes": "fff"
+        "scheduled_date":"Tue, 22 Nov 2011 06:00:00 GMT"
     }
     """
     serializer = ReminderSerializer(data=request.data)
     if serializer.is_valid():
         serialized_data = serializer.data
-        print(serialized_data)
-        message_id = serialized_data['message_id']
-        current_date = serialized_data['current_date']
-        scheduled_date = serialized_data['scheduled_date']
-        try:
-            notes_data = serialized_data['notes']
-        except:
-            notes_data = ""
-        ##calculate duration and send notification
-        local_scheduled_date = datetime.strptime(scheduled_date,'%a, %d %b %Y %H:%M:%S %Z')
+        message_id = serialized_data["message_id"]
+        current_date = serialized_data["current_date"]
+        scheduled_date = serialized_data["scheduled_date"]
+        notes_data = serialized_data["notes"]
+        # calculate duration and send notification
+        local_scheduled_date = datetime.strptime(
+            scheduled_date, "%a, %d %b %Y %H:%M:%S %Z"
+        )
         utc_scheduled_date = local_scheduled_date.replace(tzinfo=timezone.utc)
 
         local_current_date = datetime.strptime(current_date, "%a, %d %b %Y %H:%M:%S %Z")
@@ -902,14 +904,15 @@ def create_reminder(request):
                 room_id = message["room_id"]
                 try:
                     room = DB.read("dm_rooms", {"_id": room_id})
-  
+
                 except Exception as e:
                     print(e)
-                    return Response(status=status.HTTP_503_SERVICE_UNAVAILABLE) 
-                users_in_a_room = room.get("room_user_ids",[]).copy()
-                message_content = message['message']
-                sender_id = message['sender_id']
-                recipient_id = ''
+                    return Response(status=status.HTTP_503_SERVICE_UNAVAILABLE)
+
+                users_in_a_room = room.get("room_user_ids", []).copy()
+                message_content = message["message"]
+                sender_id = message["sender_id"]
+                recipient_id = ""
                 if sender_id in users_in_a_room:
                     users_in_a_room.remove(sender_id)
                     recipient_id = users_in_a_room[0]
@@ -919,7 +922,7 @@ def create_reminder(request):
                     "message": message_content,
                     "scheduled_date": scheduled_date,
                 }
-                if len(notes_data) > 0:
+                if notes_data is not None:
                     try:
                         notes = message["notes"] or []
                         notes.append(notes_data)
@@ -934,8 +937,15 @@ def create_reminder(request):
                         )
                     if response.get("status") == 200:
                         response_output["notes"] = notes
-                        return Response(data = response_output,status=status.HTTP_201_CREATED)
-                # SendNotificationThread(duration,duration_sec,utc_scheduled_date, utc_current_date).start()
+                        return Response(
+                            data=response_output, status=status.HTTP_201_CREATED
+                        )
+                SendNotificationThread(
+                    duration,
+                    duration_sec,
+                    utc_scheduled_date,
+                    utc_current_date,
+                ).start()
                 return Response(data=response_output, status=status.HTTP_201_CREATED)
             return Response(data="No such message", status=status.HTTP_400_BAD_REQUEST)
         return Response(
@@ -944,102 +954,6 @@ def create_reminder(request):
         )
     return Response(data="Bad Format ", status=status.HTTP_400_BAD_REQUEST)
 
-
-# def reminder(request):
- 
-#     # posting data to zuri core after validation
-#     plugin_id = PLUGIN_ID
-#     org_id = ORGANIZATION_ID
-#     coll_name = "reminders"
-
-#     reminder = serializer.data
-#     reminder['event_id'] = event_id
-#     reminder['user_id'] = user_id
-
-#     reminder_payload = {
-#         "plugin_id": plugin_id,
-#         "organization_id": org_id,
-#         "collection_name": coll_name,
-#         "bulk_write": False,
-#         "object_id": "",
-#         "filter": {},
-#         "payload": reminder
-#     }
-#     url = 'https://api.zuri.chat/data/write'
-
-#     try:
-#         response = requests.post(url=url, json=reminder_payload)
-
-#         if response.status_code == 201:
-#             return Response(reminder, status=status.HTTP_201_CREATED)
-#         else:
-#             return Response({"error": response.json()['message']}, status=response.status_code)
-
-#     except exceptions.ConnectionError as e:
-#         return Response(str(e), status=status.HTTP_502_BAD_GATEWAY)
-
-
-
-
-@api_view(['GET'])
-# @permission_classes((UserIsAuthenticated, ))
-def reminder_list(request):
-    """
-        This gets a list of reminders set by the user 
-        { 
-            "user_id":" "
-        }
-    """
-    if request.method == "GET":
-        # getting data from zuri core
-        DB.read("dm_messages",)
-
-        try:
-            response = requests.get(url=url)
-            if response.status_code == 200:
-                reminders_list = response.json()['data']
-                return Response(reminders_list, status=status.HTTP_200_OK)
-            else:
-                return Response({"error": response.json()["message"]}, status=response.status_code)
-        except exceptions.ConnectionError as e:
-            return Response(str(e), status=status.HTTP_502_BAD_GATEWAY)
-
-
-@api_view(['DELETE'])
-# @permission_classes((UserIsAuthenticated, ))
-def delete_reminder(request, id):
-    DB.delete(collection_name, document_id)
-
-    plugin_id = PLUGIN_ID
-    org_id = ORGANIZATION_ID
-    coll_name = "reminders"
-    if request.method == 'DELETE':
-        url = 'https://api.zuri.chat/data/delete'
-
-        payload = {
-            "plugin_id": plugin_id,
-            "organization_id": org_id,
-            "collection_name": coll_name,
-            "bulk_delete": False,
-            "object_id": id,
-            "filter": {}
-        }
-    try:
-        response = requests.post(url=url, json=payload)
-
-        if response.status_code == 200:
-            return Response({"message": "reminder successfully deleted"},
-                            status=status.HTTP_200_OK)
-        else:
-            return Response({"error": response.json()['message']}, status=response.status_code)
-
-    except exceptions.ConnectionError as e:
-        return Response(str(e), status=status.HTTP_502_BAD_GATEWAY)
-
-
-
-class Files(APIView):
-    parser_classes = (MultiPartParser, FormParser)
 
 class SendFile(APIView):
     """
